+++ packages/node/src/build.ts
@@ -9,6 +9,7 @@ import {
   resolve,
   sep,
   parse as parsePath,
+  extname,
 } from 'path';
 import { Project } from 'ts-morph';
 import { nodeFileTrace } from '@vercel/nft';
@@ -360,14 +361,19 @@ function getAWSLambdaHandler(entrypoint: string, config: Config) {
   return '';
 }
 
-export const build: BuildV3 = async ({
+export const build = async ({
   files,
   entrypoint,
+  shim,
+  useWebApi,
   workPath,
   repoRootPath,
   config = {},
   meta = {},
-}) => {
+}: Parameters<BuildV3>[0] & {
+  shim?: (handler: string) => string;
+  useWebApi?: boolean;
+}): Promise<BuildResultV3> => {
   const baseDir = repoRootPath || workPath;
   const awsLambdaHandler = getAWSLambdaHandler(entrypoint, config);
 
@@ -417,7 +423,7 @@ export const build: BuildV3 = async ({
   let routes: BuildResultV3['routes'];
   let output: BuildResultV3['output'] | undefined;
 
-  const handler = renameTStoJS(relative(baseDir, entrypointPath));
+  let handler = renameTStoJS(relative(baseDir, entrypointPath));
   const outputPath = entrypointToOutputPath(entrypoint, config.zeroConfig);
 
   // Add a `route` for Middleware
@@ -445,6 +451,30 @@ export const build: BuildV3 = async ({
     ];
   }
 
+  if (shim) {
+    const handlerFilename = basename(handler);
+    const handlerDir = dirname(handler);
+    const extension = extname(handlerFilename);
+    const extMap: Record<string, string> = {
+      '.ts': '.js',
+      '.mts': '.mjs',
+      '.mjs': '.mjs',
+      '.cjs': '.cjs',
+      '.js': '.js',
+    };
+    const ext = extMap[extension];
+    if (!ext) {
+      throw new Error(`Unsupported extension for ${entrypoint}`);
+    }
+    const filename = `shim${ext}`;
+    const shimHandler =
+      handlerDir === '.' ? filename : join(handlerDir, filename);
+    preparedFiles[shimHandler] = new FileBlob({
+      data: shim(handlerFilename),
+    });
+    handler = shimHandler;
+  }
+
   if (isEdgeFunction) {
     output = new EdgeFunction({
       entrypoint: handler,
@@ -469,7 +499,7 @@ export const build: BuildV3 = async ({
       handler,
       architecture: staticConfig?.architecture,
       runtime: nodeVersion.runtime,
-      useWebApi: isMiddleware,
+      useWebApi: isMiddleware ? true : useWebApi,
       shouldAddHelpers: isMiddleware ? false : shouldAddHelpers,
       shouldAddSourcemapSupport,
       awsLambdaHandler,